apiserver/pyproject.toml
[tool.poetry]
name = "clip-apiserver"
version = "0.0.1"
description = ""
authors = ["Parisius <paridegiunta@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
pydantic = "^2.7.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = "^2.0.29"
psycopg2-binary = "^2.9.7"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.9"
jinja2 = "^3.1.2"
black = "^24.8.0"
flake8 = "^7.1.1"
autoflake = "^2.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py311']
indent-size = 8

[tool.autoflake]
check = true